Extract visits code from maps controller
